requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.32.0
google-genai>=1.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.20
websockets>=13.0
aiofiles>=24.0.0
